requirements.txt
aiogram>=3.0.0
yfinance>=0.2.0
pandas>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24
scikit-learn>=1.3
statsmodels>=0.14
tensorflow>=2.12
matplotlib>=3.7